apps/otto-jack/requirements.txt
# Otto - Platform Guide Dependencies

# Discord Bot
discord.py==2.3.2

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client for Ollama
httpx>=0.25.0,<0.26.0

# Async support
aiofiles==23.2.1

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
